move clientchangescene from NetMan to NC
